third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-variant-serialization-expected.txt

This is a testharness.js-based test.
FAIL checking serialized value of font-variant assert_equals: expected "tabular-nums" but got "tabular-nums  "
Harness: the test ran to completion.
